requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.23.0
requests==2.32.4
pandas==2.3.1
pyodbc==5.2.0
python-dotenv==1.1.1
certifi==2025.8.3
python-dateutil==2.9.0.post0
tzdata==2025.2
azure-monitor-opentelemetry-exporter>=1.15.0
opentelemetry-sdk>=1.36.0
pytest==7.4.2